requirements.txt
absl-py==2.3.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.5
h5py==3.14.0
joblib==1.5.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
numpy==2.3.1
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rich==14.0.0
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
xgboost==3.0.2
